refactor: moving author profile provider code to provider package
