+++ packages/provider/src/author-articles-no-images.web.js
@@ -1,10 +1,10 @@
 import React from "react";
-import { AuthorArticlesNoImagesProvider } from "@times-components/provider";
+import AuthorArticlesNoImagesProvider from "./author-articles-no-images.base";
 
 export default props => (
   <AuthorArticlesNoImagesProvider
     {...props}
     shortSummaryLength={220}
-    longSummaryLength={220}
+    longSummaryLength={360}
   />
 );
