+++ packages/provider/src/provider.js
@@ -2,9 +2,8 @@ import connect from "./connect";
 import AuthorProfileProvider, {
   query as authorProfileQuery
 } from "./author-profile";
-import AuthorArticlesNoImagesProvider, {
-  query as articleListNoImagesQuery
-} from "./author-articles-no-images";
+import AuthorArticlesNoImagesProvider from "./author-articles-no-images";
+import { query as articleListNoImagesQuery } from "./author-articles-no-images.base";
 import AuthorArticlesWithImagesProvider, {
   query as articleListWithImagesQuery
 } from "./author-articles-with-images";
